pyproject.toml
[tool.poetry]
name = "lithum"
version = "0.1.0"
description = ""
authors = ["sonyakun <owner@sonyakun.com>"]
license = "GPL-3.0-or-later"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
aiodns = "3.1.1"
aiohttp = "3.9.3"
aiosignal = "1.3.1"
annotated-types = "0.6.0"
astunparse = "1.6.3"
attrs = "23.2.0"
babel = "2.14.0"
braceexpand = "0.1.7"
brotli = "1.1.0"
cachetools = "5.3.3"
cffi = "1.16.0"
click = "8.1.7"
colorama = "0.4.6"
discord-ext-help = "1.0.0"
discord-py = "2.3.2"
dnspython = "2.6.1"
enkanetwork-py = "1.4.5"
fire = "0.5.0"
frozenlist = "1.4.1"
idna = "3.6"
import-expression = "1.1.4"
iniconfig = "2.0.0"
jishaku = "2.5.2"
motor = "3.3.2"
multidict = "6.0.5"
orjson = "3.9.15"
packaging = "23.2"
pillow = "10.2.0"
pluggy = "1.4.0"
pycares = "4.4.0"
pycparser = "2.21"
pydantic = "2.6.3"
pydantic-core = "2.16.3"
pygments = "2.17.2"
pymongo = "4.6.2"
pytest = "8.0.2"
pyyaml = "6.0.1"
reactionmenu = "3.1.5"
six = "1.16.0"
termcolor = "2.4.0"
typing-extensions = "4.10.0"
yarl = "1.9.4"
dispander = {git = "https://github.com/lithum-labs/dispander"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
